Revert "Merge pull request #1244 from avgf3/cursor/investigate-and-fix-user-profile-and-room-issues-3401"
This reverts commit f8e5bb1e879ba4b4281c91120cfcc6b0ac0224a4, reversing
changes made to e15cc060eed60ee754def794447d090fffe0acf2.
