+++ server/database-setup.ts
@@ -23,7 +23,6 @@ import { ensureRoomsColumns } from './database-adapter';
 import { ensureBotsTable } from './database-adapter';
 import { ensureChatLockColumns } from './database-adapter';
 import { ensureUserPreferencesColumns } from './database-adapter';
-import { DEFAULT_ROOM_CONSTANTS } from '../client/src/utils/defaultRoomOptimizer';
 
 // إعادة تصدير دالة التهيئة من المحول
 export { initializeDatabase } from './database-adapter';
@@ -199,9 +198,9 @@ export async function createDefaultRooms(): Promise<void> {
             // اشتقاق معرف نصي ثابت للغرفة
             const derivedId =
               room.name === 'العامة'
-                ? DEFAULT_ROOM_CONSTANTS.GENERAL_ROOM_ID
+                ? 'general'
                 : room.name === 'الترحيب'
-                  ? DEFAULT_ROOM_CONSTANTS.WELCOME_ROOM_ID
+                  ? 'welcome'
                   : room.name
                       .toLowerCase()
                       .replace(/[^a-z0-9]+/gi, '-')
