+++ server/storage.ts
@@ -16,8 +16,6 @@ import {
 import { friendService } from './services/friendService';
 import { notificationService } from './services/notificationService';
 import { userService } from './services/userService';
-import { optimizedUserService } from './services/optimizedUserService';
-import { DEFAULT_ROOM_CONSTANTS } from '../client/src/utils/defaultRoomOptimizer';
 
 // Helper function
 function safeParseJsonArray(value: string): any[] {
@@ -679,13 +677,11 @@ export const storage: LegacyStorage = {
   },
 
   async setUserOnlineStatus(id: number, isOnline: boolean) {
-    // استخدام المحسن لتحديث حالة الاتصال
-    await optimizedUserService.setUserOnlineStatus(id, isOnline);
+    await databaseService.updateUser(id, { isOnline, lastSeen: new Date() });
   },
 
   async setUserCurrentRoom(id: number, currentRoom: string | null) {
-    // استخدام المحسن لتحديث الغرفة الحالية
-    await optimizedUserService.setUserCurrentRoom(id, currentRoom);
+    await databaseService.updateUser(id, { currentRoom: currentRoom || 'general', lastSeen: new Date() });
   },
 
   async createMessage(message: any) {
@@ -711,7 +707,7 @@ export const storage: LegacyStorage = {
   },
 
   async getPublicMessages(limit = 50) {
-    return await databaseService.getMessages(DEFAULT_ROOM_CONSTANTS.GENERAL_ROOM_ID, limit);
+    return await databaseService.getMessages('general', limit);
   },
 
   async getPrivateMessages(userId1: number, userId2: number, limit = 50) {
@@ -1045,12 +1041,12 @@ export const storage: LegacyStorage = {
       const room = await databaseService.getRoomById(roomId as any);
       if (room) return room as any;
 
-      // Fallback: if not found and asking for general room, synthesize minimal room info
-      if (roomId === DEFAULT_ROOM_CONSTANTS.GENERAL_ROOM_ID) {
+      // Fallback: if not found and asking for 'general', synthesize minimal room info
+      if (roomId === 'general') {
         return {
-          id: DEFAULT_ROOM_CONSTANTS.GENERAL_ROOM_ID,
-          name: DEFAULT_ROOM_CONSTANTS.GENERAL_ROOM_NAME,
-          description: DEFAULT_ROOM_CONSTANTS.DEFAULT_ROOM_DESCRIPTION,
+          id: 'general',
+          name: 'الغرفة العامة',
+          description: 'الغرفة العامة للدردشة',
           isDefault: true,
           isActive: true,
           isLocked: false,
